move STPClientAttributionMetadata
