+++ StripePayments/StripePayments/Source/API Bindings/Models/Shared/STPClientAttributionMetadata.swift
@@ -5,10 +5,10 @@
 //  Created by Joyce Qin on 7/31/25.
 //
 
-import Foundation
+@_spi(STP) import StripeCore
 
 // See https://docs.google.com/document/d/11wWdHwWzTJGe_29mHsk71fk-kG4lwvp8TLBBf4ws9JM/edit?usp=sharing
-@_spi(STP) public class STPClientAttributionMetadata: NSObject {
+@_spi(STP) public class STPClientAttributionMetadata: NSObject, Encodable, STPFormEncodable {
 
     public enum IntentCreationFlow: String {
         case standard
@@ -49,10 +49,9 @@ import Foundation
         self.paymentMethodSelectionFlow = paymentMethodSelectionFlow?.rawValue
         super.init()
     }
-}
 
-// MARK: - Encodable
-extension STPClientAttributionMetadata: Encodable {
+    
+    // MARK: - Encodable
     enum CodingKeys: String, CodingKey {
         case clientSessionId = "client_session_id"
         case elementsSessionConfigId = "elements_session_config_id"
@@ -72,5 +71,22 @@ extension STPClientAttributionMetadata: Encodable {
         try container.encode(merchantIntegrationVersion, forKey: .merchantIntegrationVersion)
         try container.encodeIfPresent(paymentIntentCreationFlow, forKey: .paymentIntentCreationFlow)
         try container.encodeIfPresent(paymentMethodSelectionFlow, forKey: .paymentMethodSelectionFlow)
+    }
+
+    // MARK: - STPFormEncodable
+    public static func rootObjectName() -> String? {
+        return "client_attribution_metadata"
+    }
+
+    public static func propertyNamesToFormFieldNamesMapping() -> [String: String] {
+        return [
+            NSStringFromSelector(#selector(getter: clientSessionId)): "client_session_id",
+            NSStringFromSelector(#selector(getter: elementsSessionConfigId)): "elements_session_config_id",
+            NSStringFromSelector(#selector(getter: merchantIntegrationSource)): "merchant_integration_source",
+            NSStringFromSelector(#selector(getter: merchantIntegrationSubtype)): "merchant_integration_subtype",
+            NSStringFromSelector(#selector(getter: merchantIntegrationVersion)): "merchant_integration_version",
+            NSStringFromSelector(#selector(getter: paymentIntentCreationFlow)): "payment_intent_creation_flow",
+            NSStringFromSelector(#selector(getter: paymentMethodSelectionFlow)): "payment_method_selection_flow",
+        ]
     }
 }
